DarkScript: Add new property, fixed repeating funcs, and other fixes
• Add new property for inheriting parents' params.
This is a very useful property, allowing subfunctions to read parents' params as local variables, just like how parents' local variables are inherited.
Note this is breaking change, so anyone testing the pre-release already will have to do the "create new and replace with another object" trick.
• Fixed the repeating events; it was running one too little. The "number of repeats" is technically 1-based but you add 1 to it for actual number of runs for a function. This seems pretty confusing, so I may amend this to a number of iterations including the original.
• Fixed aborts that ran to top level function and had their index wrap, which wasn't detected due to it being unsigned.
• Fixed Ony Any Function Aborted and Ony Any Function Completed OK not being linked.
• Made error messages in parameters/scoped variables more useful - when a variable is not found, all the scoped variables/parameters available at the function level will be written to the error message. I may make this editor-only or optional, as it's pretty verbose and also readily exposes internal variables names.

+++ DarkEdif/DarkScript/Actions.cpp
@@ -974,7 +974,7 @@ void Extension::RunningFunc_Abort(const TCHAR* error, const TCHAR* funcToUnwindT
 	if (!rf->active)
 		return CreateErrorT("Can't abort twice. Second reason discarded.");
 
-	size_t unwindToRunningFuncIndex = 0;
+	int unwindToRunningFuncIndex = 0;
 	if (funcToUnwindTo[0] != _T('\0'))
 	{
 		const std::tstring funcToUnwindToL = ToLower(funcToUnwindTo);
@@ -982,13 +982,13 @@ void Extension::RunningFunc_Abort(const TCHAR* error, const TCHAR* funcToUnwindT
 			{ return f->funcTemplate->nameL == funcToUnwindToL; });
 		if (q == globals->runningFuncs.crend())
 			return CreateErrorT("Aborting function \"%s\" to function \"%s\" failed; unwind function \"%s\" not in call stack.", rf->funcTemplate->name.c_str(), funcToUnwindTo, funcToUnwindTo);
-		unwindToRunningFuncIndex = globals->runningFuncs.size() - std::distance(globals->runningFuncs.crbegin(), q);
+		unwindToRunningFuncIndex = (int)(globals->runningFuncs.size() - std::distance(globals->runningFuncs.crbegin(), q));
 		assert((long)unwindToRunningFuncIndex >= 0);
 	}
 
-	for (size_t i = globals->runningFuncs.size() - 1; i >= unwindToRunningFuncIndex; --i)
+	for (int i = (int)globals->runningFuncs.size() - 1; i >= unwindToRunningFuncIndex; --i)
 	{
-		const auto& fa = globals->runningFuncs[i];
+		const auto& fa = globals->runningFuncs[(size_t)i];
 		fa->abortReason = error;
 		fa->currentIterationTriggering = false;
 		fa->nextRepeatIterationTriggering = false;
